responders/Gmail/requirements.txt
cortexutils~=2.1.0
google-api-python-client~=1.12.8
google-auth-httplib2~=0.0.4
google-auth-oauthlib~=0.4.2
thehive4py~=1.8.1
urllib3~=1.26.3
